refactor : 최단거리 서비스와 도메인 분리
