+++ src/main/java/wooteco/subway/admin/config/ETagHeaderFilter.java
@@ -11,7 +11,7 @@ public class ETagHeaderFilter {
 	public FilterRegistrationBean<ShallowEtagHeaderFilter> shallowEtagHeaderFilter() {
 		FilterRegistrationBean<ShallowEtagHeaderFilter> filter =
 			new FilterRegistrationBean<>(new ShallowEtagHeaderFilter());
-		filter.addUrlPatterns("/lines/detail");
+		filter.addUrlPatterns("/lines/detail", "/paths");
 		filter.setName("etagFilter");
 		return filter;
 	}
