+++ src/main/java/wooteco/subway/admin/domain/type/weightstrategy/DistanceWeightStrategy.java
@@ -1,6 +1,6 @@
 package wooteco.subway.admin.domain.type.weightstrategy;
 
-import wooteco.subway.admin.domain.LineStation;
+import wooteco.subway.admin.domain.entity.LineStation;
 
 public class DistanceWeightStrategy implements WeightStrategy {
 	@Override
