+++ src/main/java/wooteco/subway/admin/domain/type/weightstrategy/DurationWeightStrategy.java
@@ -1,6 +1,6 @@
 package wooteco.subway.admin.domain.type.weightstrategy;
 
-import wooteco.subway.admin.domain.LineStation;
+import wooteco.subway.admin.domain.entity.LineStation;
 
 public class DurationWeightStrategy implements WeightStrategy {
 	@Override
