+++ src/main/java/wooteco/subway/admin/domain/type/weightstrategy/WeightStrategy.java
@@ -1,6 +1,6 @@
 package wooteco.subway.admin.domain.type.weightstrategy;
 
-import wooteco.subway.admin.domain.LineStation;
+import wooteco.subway.admin.domain.entity.LineStation;
 
 public interface WeightStrategy {
 	double getWeight(LineStation lineStation);
