+++ src/main/java/wooteco/subway/admin/dto/LineRequest.java
@@ -1,35 +1,35 @@
 package wooteco.subway.admin.dto;
 
-import wooteco.subway.admin.domain.Line;
-
 import java.time.LocalTime;
 
+import wooteco.subway.admin.domain.entity.Line;
+
 public class LineRequest {
-    private String name;
-    private LocalTime startTime;
-    private LocalTime endTime;
-    private int intervalTime;
+	private String name;
+	private LocalTime startTime;
+	private LocalTime endTime;
+	private int intervalTime;
 
-    public LineRequest() {
-    }
+	public LineRequest() {
+	}
 
-    public String getName() {
-        return name;
-    }
+	public String getName() {
+		return name;
+	}
 
-    public LocalTime getStartTime() {
-        return startTime;
-    }
+	public LocalTime getStartTime() {
+		return startTime;
+	}
 
-    public LocalTime getEndTime() {
-        return endTime;
-    }
+	public LocalTime getEndTime() {
+		return endTime;
+	}
 
-    public int getIntervalTime() {
-        return intervalTime;
-    }
+	public int getIntervalTime() {
+		return intervalTime;
+	}
 
-    public Line toLine() {
-        return new Line(name, startTime, endTime, intervalTime);
-    }
+	public Line toLine() {
+		return new Line(name, startTime, endTime, intervalTime);
+	}
 }
