+++ src/main/java/wooteco/subway/admin/dto/StationCreateRequest.java
@@ -1,6 +1,6 @@
 package wooteco.subway.admin.dto;
 
-import wooteco.subway.admin.domain.Station;
+import wooteco.subway.admin.domain.entity.Station;
 
 public class StationCreateRequest {
 	private String name;
