+++ src/main/java/wooteco/subway/admin/repository/LineRepository.java
@@ -1,11 +1,12 @@
 package wooteco.subway.admin.repository;
 
-import org.springframework.data.repository.CrudRepository;
-import wooteco.subway.admin.domain.Line;
-
 import java.util.List;
 
+import org.springframework.data.repository.CrudRepository;
+
+import wooteco.subway.admin.domain.entity.Line;
+
 public interface LineRepository extends CrudRepository<Line, Long> {
-    @Override
-    List<Line> findAll();
+	@Override
+	List<Line> findAll();
 }
