+++ src/main/java/wooteco/subway/admin/repository/StationRepository.java
@@ -1,20 +1,21 @@
 package wooteco.subway.admin.repository;
 
-import org.springframework.data.jdbc.repository.query.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
-import wooteco.subway.admin.domain.Station;
-
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.data.jdbc.repository.query.Query;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.repository.query.Param;
+
+import wooteco.subway.admin.domain.entity.Station;
+
 public interface StationRepository extends CrudRepository<Station, Long> {
-    @Override
-    List<Station> findAllById(Iterable ids);
+	@Override
+	List<Station> findAllById(Iterable ids);
 
-    @Override
-    List<Station> findAll();
+	@Override
+	List<Station> findAll();
 
-    @Query("select * from station where name = :stationName")
-    Optional<Station> findByName(@Param("stationName") String stationName);
+	@Query("select * from station where name = :stationName")
+	Optional<Station> findByName(@Param("stationName") String stationName);
 }
