+++ src/test/java/wooteco/subway/admin/acceptance/WholeSubwayAcceptanceTest.java
@@ -6,13 +6,10 @@ import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 import wooteco.subway.admin.dto.LineDetailResponse;
 import wooteco.subway.admin.dto.LineResponse;
 import wooteco.subway.admin.dto.StationResponse;
-import wooteco.subway.admin.dto.WholeSubwayResponse;
 
 public class WholeSubwayAcceptanceTest extends AcceptanceTest {
 	@DisplayName("지하철 노선도 전체 정보 조회")
@@ -38,16 +35,5 @@ public class WholeSubwayAcceptanceTest extends AcceptanceTest {
 		assertThat(response.size()).isEqualTo(2);
 		assertThat(response.get(0).getStations().size()).isEqualTo(3);
 		assertThat(response.get(1).getStations().size()).isEqualTo(3);
-	}
-
-	private WholeSubwayResponse retrieveWholeSubway() {
-		return given().
-			accept(MediaType.APPLICATION_JSON_VALUE).
-			when().
-			get("/lines/detail").
-			then().
-			log().all().
-			statusCode(HttpStatus.OK.value()).
-			extract().as(WholeSubwayResponse.class);
 	}
 }
