+++ src/test/java/wooteco/subway/admin/api/LineControllerTest.java
@@ -18,8 +18,8 @@ import org.springframework.test.web.servlet.MvcResult;
 
 import wooteco.subway.admin.config.ETagHeaderFilter;
 import wooteco.subway.admin.controller.LineController;
-import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.domain.Station;
+import wooteco.subway.admin.domain.entity.Line;
+import wooteco.subway.admin.domain.entity.Station;
 import wooteco.subway.admin.dto.LineDetailResponse;
 import wooteco.subway.admin.dto.WholeSubwayResponse;
 import wooteco.subway.admin.service.LineService;
