+++ src/test/java/wooteco/subway/admin/repository/LineRepositoryTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
 
-import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.domain.LineStation;
+import wooteco.subway.admin.domain.entity.Line;
+import wooteco.subway.admin.domain.entity.LineStation;
 
 @DataJdbcTest
 public class LineRepositoryTest {
