+++ src/test/java/wooteco/subway/admin/repository/StationRepositoryTest.java
@@ -1,23 +1,24 @@
 package wooteco.subway.admin.repository;
 
+import static org.junit.jupiter.api.Assertions.*;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
 import org.springframework.data.relational.core.conversion.DbActionExecutionException;
-import wooteco.subway.admin.domain.Station;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import wooteco.subway.admin.domain.entity.Station;
 
 @DataJdbcTest
 public class StationRepositoryTest {
-    @Autowired
-    private StationRepository stationRepository;
+	@Autowired
+	private StationRepository stationRepository;
 
-    @Test
-    void saveStation() {
-        String stationName = "강남역";
-        stationRepository.save(new Station(stationName));
+	@Test
+	void saveStation() {
+		String stationName = "강남역";
+		stationRepository.save(new Station(stationName));
 
-        assertThrows(DbActionExecutionException.class, () -> stationRepository.save(new Station(stationName)));
-    }
+		assertThrows(DbActionExecutionException.class, () -> stationRepository.save(new Station(stationName)));
+	}
 }
